+++ json_serializable/lib/src/check_dependencies.dart
@@ -11,7 +11,7 @@ import 'package:pubspec_parse/pubspec_parse.dart';
 
 const _productionDirectories = {'lib', 'bin'};
 const _annotationPkgName = 'json_annotation';
-final requiredJsonAnnotationMinVersion = Version.parse('4.2.0');
+final requiredJsonAnnotationMinVersion = Version.parse('4.3.0');
 
 Future<void> pubspecHasRightVersion(BuildStep buildStep) async {
   final segments = buildStep.inputId.pathSegments;