+++ json_serializable/test/src/unknown_enum_value_test_input.dart
@@ -67,16 +67,6 @@ class UnknownEnumValueWrongEnumType {
 }
 
 @ShouldThrow(
-  'Error with `@JsonKey` on the `value` field. The value provided '
-  'for `unknownEnumValue` must be a matching enum.',
-)
-@JsonSerializable()
-class UnknownEnumValueNotEnumValue {
-  @JsonKey(unknownEnumValue: 'not enum value')
-  UnknownEnumValueItems? value;
-}
-
-@ShouldThrow(
   'Error with `@JsonKey` on the `value` field. `unknownEnumValue` can only be '
   'set on fields of type enum or on Iterable, List, or Set instances of an '
   'enum type.',
@@ -85,24 +75,4 @@ class UnknownEnumValueNotEnumValue {
 class UnknownEnumValueNotEnumField {
   @JsonKey(unknownEnumValue: UnknownEnumValueItems.vUnknown)
   int? value;
-}
-
-@ShouldThrow(
-  '`JsonKey.nullForUndefinedEnumValue` cannot be used with '
-  '`JsonKey.unknownEnumValue` unless the field is nullable.',
-)
-@JsonSerializable()
-class NullForUndefinedEnumValueOnNonNullableField {
-  @JsonKey(unknownEnumValue: JsonKey.nullForUndefinedEnumValue)
-  late UnknownEnumValueItems value;
-}
-
-@ShouldThrow(
-  'Error with `@JsonKey` on the `value` field. `unknownEnumValue` is '
-  '`JsonSerializable`, it must be a literal.',
-)
-@JsonSerializable()
-class WeirdValueForUnknownEnumValue {
-  @JsonKey(unknownEnumValue: JsonSerializable())
-  late UnknownEnumValueItems value;
 }